third_party/WebKit/LayoutTests/external/wpt/html/infrastructure/urls/dynamic-changes-to-base-urls/dynamic-urls.sub-expected.txt

This is a testharness.js-based test.
FAIL The 'href' attribute of the 'a' element assert_equals: The 'href' attribute is incorrect. expected "http://www.web-platform.test:8001/test.txt" but got "http://web-platform.test:8001/html/infrastructure/urls/dynamic-changes-to-base-urls/test.txt"
FAIL The 'href' attribute of the 'link' element assert_equals: The 'href' attribute is incorrect. expected "http://www.web-platform.test:8001/test.txt" but got "http://web-platform.test:8001/html/infrastructure/urls/dynamic-changes-to-base-urls/test.txt"
FAIL The 'href' attribute of the 'area' element assert_equals: The 'href' attribute is incorrect. expected "http://www.web-platform.test:8001/test.txt" but got "http://web-platform.test:8001/html/infrastructure/urls/dynamic-changes-to-base-urls/test.txt"
FAIL The 'cite' attribute of the 'q' element assert_equals: The 'cite' attribute is incorrect. expected "http://www.web-platform.test:8001/test.txt" but got "http://web-platform.test:8001/html/infrastructure/urls/dynamic-changes-to-base-urls/test.txt"
FAIL The 'cite' attribute of the 'blockquote' element assert_equals: The 'cite' attribute is incorrect. expected "http://www.web-platform.test:8001/test.txt" but got "http://web-platform.test:8001/html/infrastructure/urls/dynamic-changes-to-base-urls/test.txt"
FAIL The 'cite' attribute of the 'ins' element assert_equals: The 'cite' attribute is incorrect. expected "http://www.web-platform.test:8001/test.txt" but got "http://web-platform.test:8001/html/infrastructure/urls/dynamic-changes-to-base-urls/test.txt"
FAIL The 'cite' attribute of the 'del' element assert_equals: The 'cite' attribute is incorrect. expected "http://www.web-platform.test:8001/test.txt" but got "http://web-platform.test:8001/html/infrastructure/urls/dynamic-changes-to-base-urls/test.txt"
FAIL The 'src' attribute of the 'img' element assert_equals: The 'src' attribute is incorrect. expected "http://www.web-platform.test:8001/test.txt" but got "http://web-platform.test:8001/html/infrastructure/urls/dynamic-changes-to-base-urls/test.txt"
FAIL The 'src' attribute of the 'embed' element assert_equals: The 'src' attribute is incorrect. expected "http://www.web-platform.test:8001/test.txt" but got "http://web-platform.test:8001/html/infrastructure/urls/dynamic-changes-to-base-urls/test.txt"
FAIL The 'src' attribute of the 'video' element assert_equals: The 'src' attribute is incorrect. expected "http://www.web-platform.test:8001/test.txt" but got "http://web-platform.test:8001/html/infrastructure/urls/dynamic-changes-to-base-urls/test.txt"
FAIL The 'src' attribute of the 'iframe' element assert_equals: The 'src' attribute is incorrect. expected "http://www.web-platform.test:8001/test.txt" but got "http://web-platform.test:8001/html/infrastructure/urls/dynamic-changes-to-base-urls/test.txt"
FAIL The 'src' attribute of the 'script' element assert_equals: The 'src' attribute is incorrect. expected "http://www.web-platform.test:8001/test.txt" but got "http://web-platform.test:8001/html/infrastructure/urls/dynamic-changes-to-base-urls/test.txt"
FAIL The 'src' attribute of the 'source' element assert_equals: The 'src' attribute is incorrect. expected "http://www.web-platform.test:8001/test.txt" but got "http://web-platform.test:8001/html/infrastructure/urls/dynamic-changes-to-base-urls/test.txt"
FAIL The 'src' attribute of the 'track' element assert_equals: The 'src' attribute is incorrect. expected "http://www.web-platform.test:8001/test.txt" but got "http://web-platform.test:8001/html/infrastructure/urls/dynamic-changes-to-base-urls/test.txt"
FAIL The 'action' attribute of the 'form' element assert_equals: The 'action' attribute is incorrect. expected "http://www.web-platform.test:8001/test.txt" but got "http://web-platform.test:8001/html/infrastructure/urls/dynamic-changes-to-base-urls/test.txt"
FAIL The 'data' attribute of the 'object' element assert_equals: The 'data' attribute is incorrect. expected "http://www.web-platform.test:8001/test.txt" but got "http://web-platform.test:8001/html/infrastructure/urls/dynamic-changes-to-base-urls/test.txt"
FAIL The 'formaction' attribute of the 'button' element assert_equals: The 'formAction' attribute is incorrect. expected "http://www.web-platform.test:8001/test.txt" but got "http://web-platform.test:8001/html/infrastructure/urls/dynamic-changes-to-base-urls/test.txt"
FAIL Change the base URL must effect the descendant elements only assert_not_equals: The inner element must be effected. got disallowed value "http://web-platform.test:8001/html/infrastructure/urls/dynamic-changes-to-base-urls/test.txt"
Harness: the test ran to completion.
